server/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
python-multipart
pillow
